Don't return array, just value. This is how this API was before.
Fix Categories module default data installer.

Since the ID field of CategoryEntity is set to autogenerate, the default
ID in the default data is ignored. As a consequence the prepared data is
not saved as intended. The commit disabled autogenerate for the CategoryEntity
at install time.

Handle when manipulating root category which has a null parent (was 0 in core<=1.3.5)

Prevent data pollution in foreach loop.

Handle condition when nothing was registered, so there's nothing to delete.

Add legacy method to old D2 category entity.

Consolidate old abstract into new for BC

Add some BC methods.

Fix creation of category attributes and paths when creating a new category entry.

The path and ipath calculation was buried in the DBObject postprocess which is why it was missed.

Default new categories to active.

Add entity selector plugin

Clean up comments

+++ src/lib/util/CategoryRegistryUtil.php
@@ -39,8 +39,10 @@ class CategoryRegistryUtil
         }
 
         $entity = $em->getRepository('Zikula\Core\Doctrine\Entity\CategoryRegistryEntity')->findOneBy($params);
-        $em->remove($entity);
-        $em->flush();
+        if ($entity) {
+            $em->remove($entity);
+            $em->flush();
+        }
 
         return true;
     }
@@ -132,7 +134,7 @@ class CategoryRegistryUtil
 
         $em = \ServiceUtil::get('doctrine.entitymanager');
 
-        if (isset($catreg['id']) && is_numeric($catreg['id'])) {
+        if (isset($catreg['id'])) {
             $entity = $em->getRepository('Zikula\Core\Doctrine\Entity\CategoryRegistryEntity')->find($catreg['id']);
         } else {
             $entity = new \Zikula\Core\Doctrine\Entity\CategoryRegistryEntity;
@@ -205,7 +207,7 @@ class CategoryRegistryUtil
         /** @var $rCategory Zikula\Core\Doctrine\Entity\CategoryRegistryEntity */
         foreach ($rCategories as $rCategory) {
             $rCategory = $rCategory->toArray();
-            $fArr[$rCategory[$arraykey]] = $rCategory;
+            $fArr[$rCategory[$arraykey]] = $rCategory['category_id'];
         }
 
         $cache[$modname][$entityname] = $fArr;
